Fix windows users with space

+++ bin/cb-deltree.bat
@@ -12,16 +12,16 @@
 
 setlocal EnableDelayedExpansion
 set "CB_SILENT=false"
-if ".%1" ==".--silent" set "CB_SILENT=true" & shift
-if not exist "%1" goto END
+if ".%~1" ==".--silent" set "CB_SILENT=true" & shift
+if not exist "%~1" goto END
 
 set "TMPFILE=%TEMP%\cb-deltree-%RANDOM%%RANDOM%.tmp"
-dir /B /s "%1" > "%TMPFILE%" 2>nul
-for /f %%i in (%TMPFILE%) do (call :DELETE_FOLDER %%i)
+dir /B /s "%~1" > "%TMPFILE%" 2>nul
+for /f %%i in ("%TMPFILE%") do (call :DELETE_FOLDER "%%i")
 if exist %TMPFILE% del /f /q %TMPFILE% >nul 2>nul
-rd /s /q "%1" >nul 2>nul
+rd /s /q "%~1" >nul 2>nul
 :: repeat because rd is sometimes buggy 
-if exist "%1" rd /s /q "%1" >nul 2>nul
+if exist "%~1" rd /s /q "%1" >nul 2>nul
 goto END
 
 :DELETE_FOLDER
